[PrimitiveFitterNode] Rearrange PrimitiveFitterNode Avoid using this for the shared_pointer to avoid std::shared_with_this. The shared pointer is provided by the external object creating the instance.
